:art: Renamed auth related components and updated template following Vue style guide
