+++ web/src/views/auth/ResetConfirmation.vue
@@ -1,11 +1,13 @@
 <template>
     <div class="w-full py-8 px-12">
         <div>
-            <img alt="Bazaar"
+            <img
+                src="@/assets/img/default/trolley.png"
                 class="mx-auto h-12 w-auto"
-                src="@/assets/img/default/trolley.png" />
+                alt="Bazaar" />
 
-            <h2 class="mt-6 text-center text-3xl
+            <h2
+                class="mt-6 text-center text-3xl
                 font-bold tracking-tight text-gray-900">
                 Reset your password with Bazaar
             </h2>
@@ -16,21 +18,23 @@
                 <div class="grid grid-cols-1 gap-3">
                     <div>
                         <label class="text-sm text-gray-700">New password</label>
-                        <input type="password"
+                        <input
+                            type="password"
                             class="mt-1 block w-full rounded-lg bg-gray-50
                             border border-gray-300 text-gray-900
                             text-sm p-2.5 focus:outline-none
                             focus:shadow-outline"
                             placeholder="Enter your new password"
+                            v-model="resetForm.newPassword1"
                             :class="{
                                 'border-red-400': v$.newPassword1.$dirty &&
                                                     v$.newPassword1.$invalid 
                                 
                             }"
-                            v-model="resetForm.newPassword1"
                             @blur="v$.newPassword1.$touch" />
                         
-                        <p v-for="error of v$.newPassword1.$errors"
+                        <p
+                            v-for="error of v$.newPassword1.$errors"
                             :key="error.$uid"
                             class="mt-2 text-xs text-red-600 dark:text-red-500">
                             {{ error.$message }}
@@ -39,21 +43,23 @@
 
                     <div>
                         <label class="text-sm text-gray-700">Confirm new password</label>
-                        <input type="password"
+                        <input
+                            type="password"
                             class="mt-1 block w-full rounded-lg bg-gray-50
                             border border-gray-300 text-gray-900
                             text-sm p-2.5 focus:outline-none
                             focus:shadow-outline"
                             placeholder="Confirm your new password"
+                            v-model="resetForm.newPassword2"
                             :class="{
                                 'border-red-400': v$.newPassword2.$dirty &&
                                                     v$.newPassword2.$invalid 
                                 
                             }"
-                            v-model="resetForm.newPassword2"
                             @blur="v$.newPassword2.$touch" />
 
-                        <p v-for="error of v$.newPassword2.$errors"
+                        <p
+                            v-for="error of v$.newPassword2.$errors"
                             :key="error.$uid"
                             class="mt-2 text-xs text-red-600 dark:text-red-500">
                             {{ error.$message }}
@@ -61,7 +67,8 @@
                     </div>
 
                     <div>
-                        <button class="mt-1 group relative flex w-full justify-center
+                        <button
+                            class="mt-1 group relative flex w-full justify-center
                             rounded-lg p-2.5 border border-transparent outline-none
                             font-medium text-sm shadow-none border-solid text-white 
                             bg-green-400 border-green-400  active:bg-green-500 
@@ -83,22 +90,22 @@
 
 <script lang="ts">
 import { computed, defineComponent, onMounted, ref } from 'vue'
+import { useRoute } from 'vue-router'
 
 import type { ResetPasswordInput } from '@/common/models/auth.model'
 import { passwordRegexMedium } from '@/common/helpers'
 import router from '@/router'
 import { useAuthStore } from '@/stores'
 
+import useVuelidate from '@vuelidate/core'
 import { helpers, required, sameAs } from '@vuelidate/validators'
 import { useToast } from 'vue-toastification'
-import useVuelidate from '@vuelidate/core'
-import { useRoute } from 'vue-router'
 
 const UID_KEY = 'uid'
 const TOKEN_KEY = 'key'
 
 export default defineComponent({
-  name: 'ConfirmReset',
+  name: 'ResetConfirmation',
   setup() {
     // Form
     const resetForm = ref<ResetPasswordInput>({
@@ -144,18 +151,23 @@ export default defineComponent({
     }))
     const v$ = useVuelidate(validation, resetForm.value)
 
+    // Checkers
+    const isLoading = ref<boolean>(false)
+
     // Services
     const authStore = useAuthStore()
     const toast = useToast()
     const route = useRoute()
 
-    // Checkers
-    const isLoading = ref<boolean>(false)
-
     onMounted(() => {
         queryParamChecker()
     })
 
+    /**
+     * Check for any query parameters in path
+     * 
+     * This component requires uid and key params
+     */
     const queryParamChecker = () => {
         if (UID_KEY in route.query && TOKEN_KEY in route.query) {
             // Append query parameter value to form
@@ -167,6 +179,9 @@ export default defineComponent({
         }
     }
 
+    /**
+     * Make http request to API to confirm reset password
+     */
     const confirmReset = () => {
         isLoading.value = true
