+++ web/src/views/auth/ResetRequest.vue
@@ -1,13 +1,15 @@
 <template>
     <div class="w-full py-8 px-4 sm:px-12">
         <div>
-            <img alt="Bazaar"
+            <img
+                src="@/assets/img/default/trolley.png"
                 class="mx-auto h-12 w-auto"
-                src="@/assets/img/default/trolley.png" />
+                alt="Bazaar" />
 
-            <h2 class="mt-6 text-center text-3xl
+            <h2
+                class="mt-6 text-center text-3xl
                 font-bold tracking-tight text-gray-900">
-                Sign in to your account
+                Reset your password with Bazaar
             </h2>
         </div>
 
@@ -16,20 +18,22 @@
                 <div class="grid grid-cols-1 gap-3">
                     <div>
                         <label class="text-sm text-gray-700">Email</label>
-                        <input type="email"
+                        <input
+                            type="email"
                             class="mt-1 block w-full rounded-lg bg-gray-50
                             border border-gray-300 text-gray-900
                             text-sm p-2.5 focus:outline-none
                             focus:shadow-outline"
                             placeholder="Enter your email"
+                            v-model="resetForm.email"
                             :class="{
-                                'border-red-400': v$.username.$dirty &&
-                                                    v$.username.$invalid 
+                                'border-red-400': v$.email.$dirty &&
+                                                    v$.email.$invalid 
                                 
                             }"
-                            v-model="loginForm.username"
-                            @blur="v$.username.$touch" />
-                        <p v-for="error of v$.username.$errors"
+                            @blur="v$.email.$touch" />
+                        <p
+                            v-for="error of v$.email.$errors"
                             :key="error.$uid"
                             class="mt-2 text-xs text-red-600 dark:text-red-500">
                             {{ error.$message }}
@@ -37,30 +41,8 @@
                     </div>
 
                     <div>
-                        <label class="text-sm text-gray-700">Password</label>
-                        <input type="password"
-                            class="mt-1 block w-full rounded-lg bg-gray-50
-                            border border-gray-300 text-gray-900
-                            text-sm p-2.5 focus:outline-none
-                            focus:shadow-outline"
-                            placeholder="Enter your password"
-                            :class="{
-                                'border-red-400': v$.password.$dirty &&
-                                                    v$.password.$invalid 
-                                
-                            }"
-                            v-model="loginForm.password"
-                            @blur="v$.password.$touch" />
-
-                        <p v-for="error of v$.password.$errors"
-                            :key="error.$uid"
-                            class="mt-2 text-xs text-red-600 dark:text-red-500">
-                            {{ error.$message }}
-                        </p>
-                    </div>
-
-                    <div>
-                        <button class="mt-1 group relative flex w-full justify-center
+                        <button
+                            class="mt-1 group relative flex w-full justify-center
                             rounded-lg p-2.5 border border-transparent outline-none
                             font-medium text-sm shadow-none border-solid text-white 
                             bg-green-400 border-green-400  active:bg-green-500 
@@ -69,9 +51,9 @@
                             disabled:cursor-not-allowed focus:outline-none focus:ring-2
                             focus:ring-green-200 focus:hover:enabled:bg-green-500
                             transition-all duration-150 ease-in-out"
-                            v-on:click="login()"
+                            v-on:click="reset()"
                             :disabled="isLoading || v$.$invalid">
-                            Sign in
+                            Send request
                         </button>
                     </div>
 
@@ -82,22 +64,24 @@
                     <div class="flex flex-col">
                         <div class="text-center">
                             <p class="text-sm text-gray-500">
-                                Don't have an account?
-                                <router-link :to="{ path: '/auth/registration/customer' }"
+                                Remembered it?
+                                <router-link
+                                    :to="{ path: '/auth/login' }"
                                     class="font-medium 
                                     text-green-400 hover:text-green-300">
-                                    Sign up
+                                    Sign in
                                 </router-link>
                             </p>
                         </div>
-
+                        
                         <div class="text-center">
                             <p class="text-sm text-gray-500">
-                                Forgot your password?
-                                <router-link :to="{ path: '/auth/reset' }"
+                                Don't have an account?
+                                <router-link
+                                    :to="{ path: '/auth/registration/customer' }"
                                     class="font-medium 
                                     text-green-400 hover:text-green-300">
-                                    Reset
+                                    Sign up
                                 </router-link>
                             </p>
                         </div>
@@ -111,24 +95,22 @@
 <script lang="ts">
 import { computed, defineComponent, onMounted, ref } from 'vue'
 
-import type { LoginInput } from '@/common/models/auth.model'
-import router from '@/router'
+import type { EmailOnlyInput } from '@/common/models/auth.model'
 import { useAuthStore } from '@/stores'
 
-import { email, helpers, minLength, required } from '@vuelidate/validators'
-import { useToast } from 'vue-toastification'
 import useVuelidate from '@vuelidate/core'
+import { email, helpers, required } from '@vuelidate/validators'
+import { useToast } from 'vue-toastification'
 
 export default defineComponent({
-  name: 'Login',
+  name: 'ResetRequest',
   setup() {
     // Form
-    const loginForm = ref<LoginInput>({
-        username: null,
-        password: null
+    const resetForm = ref<EmailOnlyInput>({
+        email: null,
     })
     const validation = computed(() => ({
-        username: { 
+        email: { 
             required: helpers.withMessage(
                 'Email is required',
                 required
@@ -137,61 +119,42 @@ export default defineComponent({
                 'Valid email is required',
                 email
             )
-        },
-        password: { 
-            required: helpers.withMessage(
-                'Password is required',
-                required
-            ),
-            minLength: helpers.withMessage(
-                'Min length is 8 characters',
-                minLength(8)
-            )
         }
     }))
-    const v$ = useVuelidate(validation, loginForm.value)
+    const v$ = useVuelidate(validation, resetForm.value)
+
+    // Checkers
+    const isLoading = ref<boolean>(false)
 
     // Services
     const authStore = useAuthStore()
     const toast = useToast()
 
-    // Checkers
-    const isLoading = ref<boolean>(false)
-
     onMounted(() => {
-        // console.log('Mounted Login')
+        // console.log('Mounted ResetRequest')
     })
 
-    // Login
-    const login = () => {
+    /**
+     * Make http request to API to request reset password
+     */
+    const reset = () => {
         isLoading.value = true
 
-        return authStore.login(loginForm.value)
+        return authStore.requestReset(resetForm.value)
             .then(data => {
                 isLoading.value = false
-
+                
                 // Success toastr
-                toast.success('Success. Redirecting...')
-                
-                // Navigate home
-                const nextRoute = router.currentRoute.value.query['redirectTo'] ?
-                    String(router.currentRoute.value.query['redirectTo']) : '/home'
-                
-                router.push({ path: nextRoute })
+                toast.success('Password reset email has been sent')
             })
-            .catch(err => {
+            .catch(() => {
                 isLoading.value = false
-                
-                // If username / password error
-                if (err['status'] === 400 && 'nonFieldErrors' in err['data']) {
-                    toast.error(err['data']['nonFieldErrors'][0])
-                }
             })
     }
 
     return {
-        loginForm,
-        login,
+        resetForm,
+        reset,
         v$,
         isLoading
     }
